Bootstrap (binary) observation resampling for triptych objects

Bootstrap (binary) observation resampling assumes conditionally independent observations given the forecast value. A given number of bootstrap samples are the basis for pointwise computed confidence/consistency intervals. For every bootstrap sample, we sample observations from a Bernoulli distribution conditional on (recalibrated) forecast values.

Usage

# S3 method for triptych_murphy
resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_reliability
resampling_Bernoulli(
  x,
  level = 0.9,
  n_boot = 1000,
  position = c("diagonal", "estimate"),
  ...
)

resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_roc
resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

Arguments

x

One of the triptych objects.

level

A single value that determines which quantiles of the bootstrap sample to return. These quantiles envelop level * n_boot bootstrap draws.

n_boot

The number of bootstrap samples.

...

Additional arguments passed to other methods.

position

Either "estimate" for confidence regions, or "diagonal" for consistency regions.

Value

A list of tibbles that contain the information to draw confidence regions. The length is equal to the number of forecasting methods in x.

For the documentation of the tibble structure see murphy(), reliability(), or roc(), respectively.

Examples

data(ex_binary, package = "triptych")
tr <- triptych(ex_binary) |> dplyr::slice(1, 9)

resampling_Bernoulli(tr$murphy, n_boot = 50) |> str()
#> List of 2
#>  $ X01: tibble [1,000 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ threshold: num [1:1000] 0 0.001 0.002 0.003 0.004 ...
#>   ..$ lower    : num [1:1000] 0 0.000633 0.001197 0.001742 0.002274 ...
#>   ..$ upper    : num [1:1000] 0 0.000743 0.001389 0.002 0.002635 ...
#>   ..$ method   : chr [1:1000] "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" ...
#>   ..$ level    : num [1:1000] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
#>  $ X09: tibble [1,000 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ threshold: num [1:1000] 0 0.001 0.002 0.003 0.004 ...
#>   ..$ lower    : num [1:1000] 0 0.000997 0.001994 0.002991 0.003988 ...
#>   ..$ upper    : num [1:1000] 0 0.00112 0.00224 0.00336 0.00448 ...
#>   ..$ method   : chr [1:1000] "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" ...
#>   ..$ level    : num [1:1000] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
resampling_Bernoulli(tr$reliability, position = "estimate", n_boot = 50) |> str()
#> List of 2
#>  $ X01: tibble [998 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ x     : num [1:998] 1.19e-23 3.84e-23 9.45e-22 9.82e-20 1.13e-19 ...
#>   ..$ lower : num [1:998] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ upper : num [1:998] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ method: chr [1:998] "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" ...
#>   ..$ level : num [1:998] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
#>  $ X09: tibble [1,000 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ x     : num [1:1000] 0.00447 0.00996 0.01009 0.01073 0.012 ...
#>   ..$ lower : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ upper : num [1:1000] 0.179 0.158 0.158 0.179 0.179 ...
#>   ..$ method: chr [1:1000] "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" ...
#>   ..$ level : num [1:1000] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
resampling_Bernoulli(tr$roc, n_boot = 50) |> str()
#> List of 2
#>  $ X01: tibble [2,002 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ FAR   : num [1:2002] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ HR    : num [1:2002] 0 0.00209 0.00418 0.00628 0.00837 ...
#>   ..$ method: chr [1:2002] "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" ...
#>   ..$ level : num [1:2002] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
#>  $ X09: tibble [2,002 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ FAR   : num [1:2002] 0 0.000139 0.000292 0.000598 0.000904 ...
#>   ..$ HR    : num [1:2002] 0 0.00194 0.00386 0.00562 0.00738 ...
#>   ..$ method: chr [1:2002] "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" "resampling_Bernoulli_50" ...
#>   ..$ level : num [1:2002] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...