Bootstrap case resampling for triptych objects

Case resampling assumes independent and identically distributed forecast-observation pairs. A given number of bootstrap samples are the basis for pointwise computed confidence/consistency intervals. For every bootstrap sample, we draw forecast-observations pairs with replacement until the size of the original data set is reached.

Usage

# S3 method for triptych_murphy
resampling_cases(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_reliability
resampling_cases(x, level = 0.9, n_boot = 1000, ...)

resampling_cases(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_roc
resampling_cases(x, level = 0.9, n_boot = 1000, ...)

Arguments

x

One of the triptych objects.

level

A single value that determines which quantiles of the bootstrap sample to return. These quantiles envelop level * n_boot bootstrap draws.

n_boot

The number of bootstrap samples.

...

Additional arguments passed to other methods.

Value

A list of tibbles that contain the information to draw confidence regions. The length is equal to the number of forecasting methods in x.

For the documentation of the tibble structure see murphy(), reliability(), or roc(), respectively.

Examples

data(ex_binary, package = "triptych")
tr <- triptych(ex_binary) |> dplyr::slice(1, 9)

resampling_cases(tr$murphy, n_boot = 50) |> str()
#> List of 2
#>  $ X01: tibble [1,000 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ threshold: num [1:1000] 0 0.001 0.002 0.003 0.004 ...
#>   ..$ lower    : num [1:1000] 0 0.000641 0.001213 0.001748 0.002282 ...
#>   ..$ upper    : num [1:1000] 0 0.000737 0.001393 0.002006 0.002595 ...
#>   ..$ method   : chr [1:1000] "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" ...
#>   ..$ level    : num [1:1000] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
#>  $ X09: tibble [1,000 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ threshold: num [1:1000] 0 0.001 0.002 0.003 0.004 ...
#>   ..$ lower    : num [1:1000] 0 0.000991 0.001982 0.002973 0.003964 ...
#>   ..$ upper    : num [1:1000] 0 0.00109 0.00219 0.00328 0.00437 ...
#>   ..$ method   : chr [1:1000] "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" ...
#>   ..$ level    : num [1:1000] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
resampling_cases(tr$reliability, n_boot = 50) |> str()
#> List of 2
#>  $ X01: tibble [998 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ x     : num [1:998] 1.19e-23 3.84e-23 9.45e-22 9.82e-20 1.13e-19 ...
#>   ..$ lower : num [1:998] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ upper : num [1:998] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ method: chr [1:998] "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" ...
#>   ..$ level : num [1:998] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
#>  $ X09: tibble [1,000 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ x     : num [1:1000] 0.00447 0.00996 0.01009 0.01073 0.012 ...
#>   ..$ lower : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ upper : num [1:1000] 0 0 0 0 0.0837 ...
#>   ..$ method: chr [1:1000] "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" ...
#>   ..$ level : num [1:1000] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
resampling_cases(tr$roc, n_boot = 50) |> str()
#> List of 2
#>  $ X01: tibble [2,002 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ FAR   : num [1:2002] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ HR    : num [1:2002] 0 0.00209 0.00418 0.00628 0.00837 ...
#>   ..$ method: chr [1:2002] "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" ...
#>   ..$ level : num [1:2002] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...
#>  $ X09: tibble [2,002 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ FAR   : num [1:2002] 0 0.000316 0.000633 0.000949 0.001265 ...
#>   ..$ HR    : num [1:2002] 0 0.00175 0.00349 0.00524 0.00699 ...
#>   ..$ method: chr [1:2002] "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" "resampling_cases_50" ...
#>   ..$ level : num [1:2002] 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 0.9 ...