Bootstrap observation resampling for triptych objects

Bootstrap observation resampling for triptych objects

Usage

# S3 method for triptych_murphy
resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_reliability
resampling_Bernoulli(
  x,
  level = 0.9,
  n_boot = 1000,
  position = c("diagonal", "estimate"),
  ...
)

resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_roc
resampling_Bernoulli(x, level = 0.9, n_boot = 1000, ...)

Arguments

x

One of the triptych objects.

level

A single value that determines which quantiles of the bootstrap sample to return. These quantiles envelop level * n_boot bootstrap draws.

n_boot

The number of bootstrap samples.

...

Additional arguments passed to other methods.

position

Either "estimate" for confidence regions, or "diagonal" for consistency regions.

Value

A list of tibbles that contain the information to draw confidence regions. The length is equal to the number of forecasting methods in x.

Examples

# Triptych construction
forecasts <- matrix(runif(300), ncol = 3)
colnames(forecasts) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, forecasts[, 1])
tr <- triptych(forecasts, observations)

resampling_cases(tr$murphy, n_boot = 50)
#> $Method_1
#> # A tibble: 1,000 × 5
#>    threshold    lower   upper method              level
#>        <dbl>    <dbl>   <dbl> <chr>               <dbl>
#>  1   0       0        0       resampling_cases_50   0.9
#>  2   0.00100 0.000841 0.00124 resampling_cases_50   0.9
#>  3   0.00200 0.00168  0.00248 resampling_cases_50   0.9
#>  4   0.00300 0.00252  0.00372 resampling_cases_50   0.9
#>  5   0.00400 0.00336  0.00496 resampling_cases_50   0.9
#>  6   0.00501 0.00420  0.00621 resampling_cases_50   0.9
#>  7   0.00601 0.00505  0.00745 resampling_cases_50   0.9
#>  8   0.00701 0.00589  0.00869 resampling_cases_50   0.9
#>  9   0.00801 0.00673  0.00993 resampling_cases_50   0.9
#> 10   0.00901 0.00757  0.0112  resampling_cases_50   0.9
#> # ℹ 990 more rows
#> 
#> $Method_2
#> # A tibble: 1,000 × 5
#>    threshold    lower   upper method              level
#>        <dbl>    <dbl>   <dbl> <chr>               <dbl>
#>  1   0       0        0       resampling_cases_50   0.9
#>  2   0.00100 0.000821 0.00116 resampling_cases_50   0.9
#>  3   0.00200 0.00164  0.00232 resampling_cases_50   0.9
#>  4   0.00300 0.00246  0.00348 resampling_cases_50   0.9
#>  5   0.00400 0.00328  0.00464 resampling_cases_50   0.9
#>  6   0.00501 0.00410  0.00581 resampling_cases_50   0.9
#>  7   0.00601 0.00492  0.00697 resampling_cases_50   0.9
#>  8   0.00701 0.00575  0.00813 resampling_cases_50   0.9
#>  9   0.00801 0.00657  0.00929 resampling_cases_50   0.9
#> 10   0.00901 0.00739  0.0105  resampling_cases_50   0.9
#> # ℹ 990 more rows
#> 
#> $Method_3
#> # A tibble: 1,000 × 5
#>    threshold    lower   upper method              level
#>        <dbl>    <dbl>   <dbl> <chr>               <dbl>
#>  1   0       0        0       resampling_cases_50   0.9
#>  2   0.00100 0.000841 0.00120 resampling_cases_50   0.9
#>  3   0.00200 0.00168  0.00240 resampling_cases_50   0.9
#>  4   0.00300 0.00252  0.00360 resampling_cases_50   0.9
#>  5   0.00400 0.00336  0.00480 resampling_cases_50   0.9
#>  6   0.00501 0.00420  0.00601 resampling_cases_50   0.9
#>  7   0.00601 0.00505  0.00721 resampling_cases_50   0.9
#>  8   0.00701 0.00589  0.00841 resampling_cases_50   0.9
#>  9   0.00801 0.00673  0.00961 resampling_cases_50   0.9
#> 10   0.00901 0.00757  0.0108  resampling_cases_50   0.9
#> # ℹ 990 more rows
#> 
resampling_cases(tr$reliability, position = "estimate", n_boot = 50)
#> $Method_1
#> # A tibble: 100 × 5
#>         x lower upper method              level
#>     <dbl> <dbl> <dbl> <chr>               <dbl>
#>  1 0.0177     0     0 resampling_cases_50   0.9
#>  2 0.0322     0     0 resampling_cases_50   0.9
#>  3 0.0389     0     0 resampling_cases_50   0.9
#>  4 0.0653     0     0 resampling_cases_50   0.9
#>  5 0.0795     0     0 resampling_cases_50   0.9
#>  6 0.118      0     0 resampling_cases_50   0.9
#>  7 0.118      0     0 resampling_cases_50   0.9
#>  8 0.130      0     0 resampling_cases_50   0.9
#>  9 0.132      0     0 resampling_cases_50   0.9
#> 10 0.133      0     0 resampling_cases_50   0.9
#> # ℹ 90 more rows
#> 
#> $Method_2
#> # A tibble: 100 × 5
#>         x lower upper method              level
#>     <dbl> <dbl> <dbl> <chr>               <dbl>
#>  1 0.0187 0     0     resampling_cases_50   0.9
#>  2 0.0205 0     0.147 resampling_cases_50   0.9
#>  3 0.0246 0     0.519 resampling_cases_50   0.9
#>  4 0.0326 0     0.519 resampling_cases_50   0.9
#>  5 0.0354 0     0.545 resampling_cases_50   0.9
#>  6 0.0502 0     0.551 resampling_cases_50   0.9
#>  7 0.0555 0     0.551 resampling_cases_50   0.9
#>  8 0.0579 0.286 0.551 resampling_cases_50   0.9
#>  9 0.0661 0.286 0.551 resampling_cases_50   0.9
#> 10 0.0825 0.286 0.551 resampling_cases_50   0.9
#> # ℹ 90 more rows
#> 
#> $Method_3
#> # A tibble: 100 × 5
#>         x lower upper method              level
#>     <dbl> <dbl> <dbl> <chr>               <dbl>
#>  1 0.0716 0     0     resampling_cases_50   0.9
#>  2 0.0771 0     0.182 resampling_cases_50   0.9
#>  3 0.0803 0     0.526 resampling_cases_50   0.9
#>  4 0.0840 0     0.526 resampling_cases_50   0.9
#>  5 0.0980 0     0.526 resampling_cases_50   0.9
#>  6 0.114  0.1   0.526 resampling_cases_50   0.9
#>  7 0.132  0.1   0.526 resampling_cases_50   0.9
#>  8 0.139  0.125 0.526 resampling_cases_50   0.9
#>  9 0.144  0.125 0.526 resampling_cases_50   0.9
#> 10 0.169  0.125 0.526 resampling_cases_50   0.9
#> # ℹ 90 more rows
#> 
resampling_cases(tr$roc, n_boot = 50)
#> $Method_1
#> # A tibble: 202 × 4
#>        FAR     HR method              level
#>      <dbl>  <dbl> <chr>               <dbl>
#>  1 0       0      resampling_cases_50   0.9
#>  2 0.00127 0.0187 resampling_cases_50   0.9
#>  3 0.00310 0.0369 resampling_cases_50   0.9
#>  4 0.00465 0.0554 resampling_cases_50   0.9
#>  5 0.00624 0.0738 resampling_cases_50   0.9
#>  6 0.00817 0.0918 resampling_cases_50   0.9
#>  7 0.0110  0.109  resampling_cases_50   0.9
#>  8 0.0141  0.126  resampling_cases_50   0.9
#>  9 0.0174  0.143  resampling_cases_50   0.9
#> 10 0.0207  0.159  resampling_cases_50   0.9
#> # ℹ 192 more rows
#> 
#> $Method_2
#> # A tibble: 202 × 4
#>        FAR     HR method              level
#>      <dbl>  <dbl> <chr>               <dbl>
#>  1 0       0      resampling_cases_50   0.9
#>  2 0.00989 0.0101 resampling_cases_50   0.9
#>  3 0.0198  0.0202 resampling_cases_50   0.9
#>  4 0.0297  0.0303 resampling_cases_50   0.9
#>  5 0.0395  0.0405 resampling_cases_50   0.9
#>  6 0.0494  0.0506 resampling_cases_50   0.9
#>  7 0.0593  0.0607 resampling_cases_50   0.9
#>  8 0.0692  0.0708 resampling_cases_50   0.9
#>  9 0.0791  0.0809 resampling_cases_50   0.9
#> 10 0.0890  0.0910 resampling_cases_50   0.9
#> # ℹ 192 more rows
#> 
#> $Method_3
#> # A tibble: 202 × 4
#>        FAR     HR method              level
#>      <dbl>  <dbl> <chr>               <dbl>
#>  1 0       0      resampling_cases_50   0.9
#>  2 0.00941 0.0106 resampling_cases_50   0.9
#>  3 0.0188  0.0212 resampling_cases_50   0.9
#>  4 0.0282  0.0318 resampling_cases_50   0.9
#>  5 0.0376  0.0424 resampling_cases_50   0.9
#>  6 0.0471  0.0529 resampling_cases_50   0.9
#>  7 0.0565  0.0635 resampling_cases_50   0.9
#>  8 0.0659  0.0741 resampling_cases_50   0.9
#>  9 0.0753  0.0847 resampling_cases_50   0.9
#> 10 0.0848  0.0952 resampling_cases_50   0.9
#> # ℹ 192 more rows
#>