Bootstrap case resampling for triptych objects

Bootstrap case resampling for triptych objects

Usage

# S3 method for triptych_murphy
resampling_cases(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_reliability
resampling_cases(x, level = 0.9, n_boot = 1000, ...)

resampling_cases(x, level = 0.9, n_boot = 1000, ...)

# S3 method for triptych_roc
resampling_cases(x, level = 0.9, n_boot = 1000, ...)

Arguments

x

One of the triptych objects.

level

A single value that determines which quantiles of the bootstrap sample to return. These quantiles envelop level * n_boot bootstrap draws.

n_boot

The number of bootstrap samples.

...

Additional arguments passed to other methods.

Value

A list of tibbles that contain the information to draw confidence regions. The length is equal to the number of forecasting methods in x.

Examples

# Triptych construction
forecasts <- matrix(runif(300), ncol = 3)
colnames(forecasts) <- c("Method_1", "Method_2", "Method_3")
observations <- rbinom(100, 1, forecasts[, 1])
tr <- triptych(forecasts, observations)

resampling_cases(tr$murphy, n_boot = 50)
#> $Method_1
#> # A tibble: 1,000 × 5
#>    threshold    lower   upper method              level
#>        <dbl>    <dbl>   <dbl> <chr>               <dbl>
#>  1   0       0        0       resampling_cases_50   0.9
#>  2   0.00100 0.000761 0.00102 resampling_cases_50   0.9
#>  3   0.00200 0.00152  0.00204 resampling_cases_50   0.9
#>  4   0.00300 0.00228  0.00306 resampling_cases_50   0.9
#>  5   0.00400 0.00304  0.00408 resampling_cases_50   0.9
#>  6   0.00501 0.00380  0.00511 resampling_cases_50   0.9
#>  7   0.00601 0.00456  0.00613 resampling_cases_50   0.9
#>  8   0.00701 0.00519  0.00715 resampling_cases_50   0.9
#>  9   0.00801 0.00593  0.00817 resampling_cases_50   0.9
#> 10   0.00901 0.00667  0.00919 resampling_cases_50   0.9
#> # ℹ 990 more rows
#> 
#> $Method_2
#> # A tibble: 1,000 × 5
#>    threshold    lower   upper method              level
#>        <dbl>    <dbl>   <dbl> <chr>               <dbl>
#>  1   0       0        0       resampling_cases_50   0.9
#>  2   0.00100 0.000701 0.00106 resampling_cases_50   0.9
#>  3   0.00200 0.00140  0.00212 resampling_cases_50   0.9
#>  4   0.00300 0.00210  0.00318 resampling_cases_50   0.9
#>  5   0.00400 0.00280  0.00424 resampling_cases_50   0.9
#>  6   0.00501 0.00350  0.00531 resampling_cases_50   0.9
#>  7   0.00601 0.00420  0.00637 resampling_cases_50   0.9
#>  8   0.00701 0.00490  0.00743 resampling_cases_50   0.9
#>  9   0.00801 0.00561  0.00849 resampling_cases_50   0.9
#> 10   0.00901 0.00631  0.00955 resampling_cases_50   0.9
#> # ℹ 990 more rows
#> 
#> $Method_3
#> # A tibble: 1,000 × 5
#>    threshold    lower   upper method              level
#>        <dbl>    <dbl>   <dbl> <chr>               <dbl>
#>  1   0       0        0       resampling_cases_50   0.9
#>  2   0.00100 0.000741 0.00102 resampling_cases_50   0.9
#>  3   0.00200 0.00160  0.0616  resampling_cases_50   0.9
#>  4   0.00300 0.00240  0.0624  resampling_cases_50   0.9
#>  5   0.00400 0.00320  0.0632  resampling_cases_50   0.9
#>  6   0.00501 0.00400  0.0640  resampling_cases_50   0.9
#>  7   0.00601 0.00480  0.0648  resampling_cases_50   0.9
#>  8   0.00701 0.00561  0.0656  resampling_cases_50   0.9
#>  9   0.00801 0.00641  0.0664  resampling_cases_50   0.9
#> 10   0.00901 0.00721  0.0672  resampling_cases_50   0.9
#> # ℹ 990 more rows
#> 
resampling_cases(tr$reliability, n_boot = 50)
#> $Method_1
#> # A tibble: 100 × 5
#>          x lower  upper method              level
#>      <dbl> <dbl>  <dbl> <chr>               <dbl>
#>  1 0.00650     0 0      resampling_cases_50   0.9
#>  2 0.0289      0 0      resampling_cases_50   0.9
#>  3 0.0324      0 0      resampling_cases_50   0.9
#>  4 0.0331      0 0      resampling_cases_50   0.9
#>  5 0.0528      0 0      resampling_cases_50   0.9
#>  6 0.0623      0 0      resampling_cases_50   0.9
#>  7 0.0675      0 0.0207 resampling_cases_50   0.9
#>  8 0.129       0 0.185  resampling_cases_50   0.9
#>  9 0.140       0 0.185  resampling_cases_50   0.9
#> 10 0.156       0 0.185  resampling_cases_50   0.9
#> # ℹ 90 more rows
#> 
#> $Method_2
#> # A tibble: 100 × 5
#>         x  lower upper method              level
#>     <dbl>  <dbl> <dbl> <chr>               <dbl>
#>  1 0.0135 0.167  0.6   resampling_cases_50   0.9
#>  2 0.0162 0      0.582 resampling_cases_50   0.9
#>  3 0.0192 0      0.582 resampling_cases_50   0.9
#>  4 0.0202 0.0512 0.6   resampling_cases_50   0.9
#>  5 0.0604 0.25   0.6   resampling_cases_50   0.9
#>  6 0.0685 0.25   0.6   resampling_cases_50   0.9
#>  7 0.0697 0.25   0.606 resampling_cases_50   0.9
#>  8 0.0869 0.25   0.606 resampling_cases_50   0.9
#>  9 0.0961 0.272  0.606 resampling_cases_50   0.9
#> 10 0.106  0.424  0.606 resampling_cases_50   0.9
#> # ℹ 90 more rows
#> 
#> $Method_3
#> # A tibble: 100 × 5
#>          x lower upper method              level
#>      <dbl> <dbl> <dbl> <chr>               <dbl>
#>  1 0.00167 0.355 0.586 resampling_cases_50   0.9
#>  2 0.00946 0.355 0.569 resampling_cases_50   0.9
#>  3 0.0200  0.312 0.569 resampling_cases_50   0.9
#>  4 0.0246  0     0.569 resampling_cases_50   0.9
#>  5 0.0355  0.125 0.569 resampling_cases_50   0.9
#>  6 0.0466  0.125 0.569 resampling_cases_50   0.9
#>  7 0.0515  0.188 0.586 resampling_cases_50   0.9
#>  8 0.0516  0.188 0.586 resampling_cases_50   0.9
#>  9 0.0660  0.188 0.586 resampling_cases_50   0.9
#> 10 0.0776  0.241 0.586 resampling_cases_50   0.9
#> # ℹ 90 more rows
#> 
resampling_cases(tr$roc, n_boot = 50)
#> $Method_1
#> # A tibble: 202 × 4
#>      FAR     HR method              level
#>    <dbl>  <dbl> <chr>               <dbl>
#>  1     0 0      resampling_cases_50   0.9
#>  2     0 0.0182 resampling_cases_50   0.9
#>  3     0 0.0364 resampling_cases_50   0.9
#>  4     0 0.0545 resampling_cases_50   0.9
#>  5     0 0.0727 resampling_cases_50   0.9
#>  6     0 0.0909 resampling_cases_50   0.9
#>  7     0 0.109  resampling_cases_50   0.9
#>  8     0 0.127  resampling_cases_50   0.9
#>  9     0 0.145  resampling_cases_50   0.9
#> 10     0 0.164  resampling_cases_50   0.9
#> # ℹ 192 more rows
#> 
#> $Method_2
#> # A tibble: 202 × 4
#>        FAR     HR method              level
#>      <dbl>  <dbl> <chr>               <dbl>
#>  1 0       0      resampling_cases_50   0.9
#>  2 0.00805 0.0116 resampling_cases_50   0.9
#>  3 0.0161  0.0232 resampling_cases_50   0.9
#>  4 0.0241  0.0348 resampling_cases_50   0.9
#>  5 0.0322  0.0464 resampling_cases_50   0.9
#>  6 0.0402  0.0580 resampling_cases_50   0.9
#>  7 0.0503  0.0679 resampling_cases_50   0.9
#>  8 0.0604  0.0778 resampling_cases_50   0.9
#>  9 0.0705  0.0877 resampling_cases_50   0.9
#> 10 0.0806  0.0977 resampling_cases_50   0.9
#> # ℹ 192 more rows
#> 
#> $Method_3
#> # A tibble: 202 × 4
#>        FAR     HR method              level
#>      <dbl>  <dbl> <chr>               <dbl>
#>  1 0       0      resampling_cases_50   0.9
#>  2 0.00939 0.0105 resampling_cases_50   0.9
#>  3 0.0188  0.0210 resampling_cases_50   0.9
#>  4 0.0282  0.0315 resampling_cases_50   0.9
#>  5 0.0375  0.0420 resampling_cases_50   0.9
#>  6 0.0469  0.0525 resampling_cases_50   0.9
#>  7 0.0563  0.0630 resampling_cases_50   0.9
#>  8 0.0657  0.0735 resampling_cases_50   0.9
#>  9 0.0751  0.0840 resampling_cases_50   0.9
#> 10 0.0845  0.0945 resampling_cases_50   0.9
#> # ℹ 192 more rows
#>